Updated application name to QuikDash, changed all file names and deleted unused files
